+++ packages/frontend/src/components/UserSettings/SocialLoginsPanel/index.tsx
@@ -33,6 +33,8 @@ const isIssuerTypeAvailable = (
             return health.auth.azuread.enabled;
         case OpenIdIdentityIssuerType.GENERIC_OIDC:
             return health.auth.oidc.enabled;
+        case OpenIdIdentityIssuerType.SNOWFLAKE:
+            return health.auth.snowflake.enabled;
         default:
             return assertUnreachable(
                 issuerType,